pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Johnny32id <johnnyhandsome3232@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^22.0.0"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.28.3"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.12.0"
pytest-flask = "^1.3.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.coverage.run]
source = ["page_analyzer"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
